pyproject.toml
[tool.poetry]
name = "draft"
version = "0.1.0"
description = ""
authors = ["Francesco Frassinelli <fraph24@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mercantile = "^1.1.6"
pypng = "^0.0.20"
aiohttp = "^3.7.4"
install = "^1.3.4"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]
black = "^21.4b2"
flakehell = "^0.9.0"
isort = "^5.8.0"
pre-commit = "^2.12.1"

[tool.flakehell]
# https://flakehell.readthedocs.io/
# https://black.readthedocs.io/en/stable/compatible_configs.html?highlight=flake8#flake8
max-line-length = 88
extend-ignore = "E203, W503"
# https://github.com/flakehell/flakehell/issues/10#issuecomment-822720074
extended_default_ignore = [] 

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility/
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
